+++ code/src/user_interface.h
@@ -36,7 +36,7 @@ public:
     virtual void switcher(const uint8_t index, const uint8_t count, const char *options[]) = 0;
     virtual void recipeSummary(const char *name, const char *description) = 0;
     virtual void recipeConfigCoffeeWeight(const char *header, unsigned int weightMg, unsigned int waterWeightMl) = 0;
-    virtual void recipeConfigRatio(const char *header, float coffee, float water) = 0;
+    virtual void recipeConfigRatio(const char *header, uint32_t coffee, uint32_t water) = 0;
     virtual void recipeInsertCoffee(int32_t weightMg, uint32_t requiredWeightMg) = 0;
     virtual void recipePour(const char *text, int32_t weightToPourMg, uint64_t timeToFinishMs, bool isPause, uint8_t pourIndex, uint8_t pours) = 0;
     virtual void text(const char *text) = 0;
